+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix duplicate recipients in "To" and "Cc" on reply (#8912)
 - Fix bug where it wasn't possible to scroll lists by clicking middle mouse button (#8942)
 - Fix bug where label text in a single-input dialog could be partially invisible in some locales (#8905)
+- Fix bug where LDAP (fulltext) search didn't work without 'search_fields' in config (#8874)
 
 ## Release 1.6.1